Split most of index.ts into smaller files
Fixes #89. I hope you won't think I'm too forward with this change ☹ it was taking a while to grok through `index.ts` and this helped me understand it quite a bit.

These are the only real code changes I made other to pass through parameters, and for each it was only to improve clarity:
* Rename `computeKey` to `computeRuleFailureKey`
* Rename `filterProblemsForDocument` to `filterFailuresForDocument` _(and generally consolidated "problem" into "failure")_
