+++ src/codes.ts
@@ -1,0 +1,1 @@
+export const TSLINT_ERROR_CODE = 100000;
