+++ src/fixes/addDisableRuleFix.ts
@@ -1,0 +1,15 @@
+import * as tslint_module from 'tslint';
+import * as ts_module from 'typescript';
+
+export function addDisableRuleFix(fixes: ts_module.CodeAction[], problem: tslint_module.RuleFailure, fileName: string, file: ts_module.SourceFile) {
+    fixes.push({
+        description: `Disable rule '${problem.getRuleName()}'`,
+        changes: [{
+            fileName: fileName,
+            textChanges: [{
+                newText: `// tslint:disable-next-line:${problem.getRuleName()}\n`,
+                span: { start: file.getLineStarts()[problem.getStartPosition().getLineAndCharacter().line], length: 0 }
+            }]
+        }]
+    });
+}
